+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Display orchestrator and chaincode versions in node info command
+- Accept full datamanager, metrics and parent_tasks objects in get_*_tuple responses
 
 ## [0.11.0] - 2021-10-04