+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
@@ -7,15 +8,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+
 - Display orchestrator and chaincode versions in node info command
 - Accept full datamanager, metrics and parent_tasks objects in get_*_tuple responses
+- Add task extra information related to start_date and end_date
 
 ## [0.11.0] - 2021-10-04
 
 ### Added
+
 - Display backend version in node info command
 
 ### Changed
+
 - Internal: backend now serves paginated lists of asset. Client still returns simple lists as before.
 - [BREAKING CHANGE] All algos are now one single object with a category property.
 - [BREAKING CHANGE] Various API changes due to the architecture change introduced by the orchestrator.